New file distributed.py
