+++ genrl/agents/deep/a2c/a2c.py
@@ -6,7 +6,12 @@ import torch.optim as opt
 from torch.nn import functional as F
 
 from genrl.agents.deep.base import OnPolicyAgent
-from genrl.utils import get_env_properties, get_model, safe_mean
+from genrl.utils import (
+    compute_returns_and_advantage,
+    get_env_properties,
+    get_model,
+    safe_mean,
+)
 
 
 class A2C(OnPolicyAgent):
@@ -123,8 +128,8 @@ class A2C(OnPolicyAgent):
             values (:obj:`torch.Tensor`): Values of states encountered during the rollout
             dones (:obj:`list` of bool): Game over statuses of each environment
         """
-        self.compute_returns_and_advantage(
-            values.detach().cpu().numpy(), dones.cpu().numpy()
+        compute_returns_and_advantage(
+            self.rollout, values.detach().cpu().numpy(), dones.cpu().numpy()
         )
 
     def evaluate_actions(self, states: torch.Tensor, actions: torch.Tensor):
