+++ genrl/agents/deep/base/onpolicy.py
@@ -95,45 +95,3 @@ class OnPolicyAgent(BaseAgent):
             self.collect_rewards(dones, i)
 
         return values, dones
-
-    def compute_returns_and_advantage(
-        self,
-        last_value: torch.Tensor,
-        dones: np.ndarray,
-        use_gae: bool = False,
-    ) -> None:
-        """
-        Post-processing function: compute the returns (sum of discounted rewards)
-        and advantage (A(s) = R - V(S)).
-        Adapted from Stable-Baselines PPO2.
-        ;param rollout_buffer: (RolloutBuffer, BaseBuffer) An instance of the rollout buffer used in On-policy algorithms
-        :param last_value: (torch.Tensor)
-        :param dones: (np.ndarray)
-        :param use_gae: (bool) Whether to use Generalized Advantage Estimation
-            or normal advantage for advantage computation.
-        """
-        last_value = last_value.flatten()
-
-        if use_gae:
-            gae_lambda = self.rollout.gae_lambda
-        else:
-            gae_lambda = 1
-
-        next_values = last_value
-        next_non_terminal = 1 - dones
-
-        running_advantage = 0.0
-        for step in reversed(range(self.rollout.buffer_size)):
-            delta = (
-                self.rollout.rewards[step]
-                + self.rollout.gamma * next_non_terminal * next_values
-                - self.rollout.values[step]
-            )
-            running_advantage = (
-                delta + self.rollout.gamma * gae_lambda * running_advantage
-            )
-            next_non_terminal = 1 - self.rollout.dones[step]
-            next_values = self.rollout.values[step]
-            self.rollout.advantages[step] = running_advantage
-
-        self.rollout.returns = self.rollout.advantages + self.rollout.values
