+++ genrl/agents/deep/vpg/vpg.py
@@ -5,7 +5,12 @@ import torch
 import torch.optim as opt
 
 from genrl.agents import OnPolicyAgent
-from genrl.utils import get_env_properties, get_model, safe_mean
+from genrl.utils import (
+    compute_returns_and_advantage,
+    get_env_properties,
+    get_model,
+    safe_mean,
+)
 
 
 class VPG(OnPolicyAgent):
@@ -111,8 +116,8 @@ class VPG(OnPolicyAgent):
             values (:obj:`torch.Tensor`): Values of states encountered during the rollout
             dones (:obj:`list` of bool): Game over statuses of each environment
         """
-        self.compute_returns_and_advantage(
-            values.detach().cpu().numpy(), dones.cpu().numpy()
+        compute_returns_and_advantage(
+            self.rollout, values.detach().cpu().numpy(), dones.cpu().numpy()
         )
 
     def update_params(self) -> None:
