+++ genrl/utils/__init__.py
@@ -7,6 +7,7 @@ from genrl.utils.data_bandits import (  # noqa
     MushroomDataBandit,
     StatlogDataBandit,
 )
+from genrl.utils.discount import compute_returns_and_advantage  # noqa
 from genrl.utils.logger import CSVLogger  # noqa
 from genrl.utils.logger import HumanOutputFormat  # noqa
 from genrl.utils.logger import Logger  # noqa
